dada

+++ frontend/src/components/admin/adminServices/deleteService/DeleteService.jsx
@@ -7,24 +7,30 @@ const DeleteService = ({categorizedServices, setCategorizedServices}) => {
 
 	const selectedService = useSelector((state) => state.admin.selectedService)
 	const delService = () => {
-		axios.delete('http://188.225.36.185/api/service/delete-service/', {
-			data: {id: selectedService.slice(1)},
-			headers: {Authorization: `Bearer ${localStorage.getItem('token')}`}
-		})
-		.then(() => {
-			setCategorizedServices((prevCategorizedServices) => {
-				return prevCategorizedServices.map((category) => {
-					return {
-						...category,
-						services: category.services.filter((service) => service.id !== Number(selectedService.slice(1)))
-					};
+		if(selectedService) {
+			axios.delete('http://188.225.36.185/api/service/delete-service/', {
+				data: {id: selectedService.slice(1)},
+				headers: {Authorization: `Bearer ${localStorage.getItem('token')}`}
+			})
+			.then(() => {
+				document.querySelector('.del_btn').style.border = '2px solid green'
+				setCategorizedServices((prevCategorizedServices) => {
+					return prevCategorizedServices.map((category) => {
+						return {
+							...category,
+							services: category.services.filter((service) => service.id !== Number(selectedService.slice(1)))
+						};
+					});
 				});
-			});
-		})
-		.catch((error) => {
-			//TODO: сделать стили при ошибки
-			console.error('Ошибка при удалении услуги', error)
-		})
+			})
+			.catch((error) => {
+				console.error('Ошибка при удалении услуги', error)
+				document.querySelector('.del_btn').style.border = '2px solid red'
+			})
+		} 
+		else{
+			document.querySelector('.del_btn').style.border = '2px solid red'
+		}
 	}
 
 	return (
@@ -36,7 +42,7 @@ const DeleteService = ({categorizedServices, setCategorizedServices}) => {
 			<div className="deleteService_field">
 				{categorizedServices.map(category => <ServiceList key={category.category_id} category={category} />)}
 			</div>
-			<input type="submit" value='Удалить' onClick={() => delService()}/>
+			<input className='del_btn' type="submit" value='Удалить' onClick={() => delService()}/>
 		</div>
 	)
 }
